Download delete optimization - navigate to home once all tours deleted

+++ app/src/main/java/com/ithakatales/android/download/TourDownloader.java
@@ -91,9 +91,15 @@ public class TourDownloader {
         attractionRepo.remove(attraction.getId());
     }
 
-    public void deleteAll() {
-        tourStorage.removeAllTours();
-        attractionRepo.removeAll();
+    public boolean deleteAll() {
+        boolean isEmpty = attractionRepo.readAll().isEmpty();
+
+        if ( ! isEmpty) {
+            tourStorage.removeAllTours();
+            attractionRepo.removeAll();
+        }
+
+        return ! isEmpty;
     }
 
     public TourDownloadProgress readProgress(long attractionId) {
